ra_salsa
Struct QueryTableMut

pub struct QueryTableMut<'me, Q>
where
    Q: Query + 'me,
{ /* private fields */ }

Return value from the query_mut method on Database. Gives access to the set method, notably, that is used to set the value of an input query.

Implementations

impl<'me, Q> QueryTableMut<'me, Q>
where Q: Query,

pub fn new(runtime: &'me mut Runtime, storage: &'me Q::Storage) -> Self

Constructs a new QueryTableMut.

pub fn set(&mut self, key: Q::Key, value: Q::Value)
where Q::Storage: InputQueryStorageOps<Q>,

Assign a value to an “input query”. Must be used outside of an active query computation.

If you are using snapshot, see the notes on blocking and cancellation on the query_mut method.

pub fn set_with_durability( &mut self, key: Q::Key, value: Q::Value, durability: Durability, )
where Q::Storage: InputQueryStorageOps<Q>,

Assign a value to an “input query”, with the additional promise that this value will never change. Must be used outside of an active query computation.

If you are using snapshot, see the notes on blocking and cancellation on the query_mut method.

pub fn set_lru_capacity(&self, cap: u16)
where Q::Storage: LruQueryStorageOps,

Sets the size of LRU cache of values for this query table.

That is, at most cap values will be preset in the table at the same time. This helps with keeping maximum memory usage under control, at the cost of potential extra recalculations of evicted values.

If cap is zero, all values are preserved, this is the default.

pub fn invalidate(&mut self, key: &Q::Key)
where Q::Storage: DerivedQueryStorageOps<Q>,

Marks the computed value as outdated.

This causes salsa to re-execute the query function on the next access to the query, even if all dependencies are up to date.

This is most commonly used as part of the on-demand input pattern.