ra_salsa
Trait InternValue

pub trait InternValue {
    type Key: Eq + Hash + Debug + Clone;

    // Required method
    fn into_key(&self) -> Self::Key;

    // Provided method
    fn with_key<F: FnOnce(&Self::Key) -> T, T>(&self, f: F) -> T { ... }
}

Trait implemented for the “value” that is being interned.

Required Associated Types

type Key: Eq + Hash + Debug + Clone

They key used to intern this value by.

Required Methods

fn into_key(&self) -> Self::Key

Maps the value to a key that will be used to intern it.

Provided Methods

fn with_key<F: FnOnce(&Self::Key) -> T, T>(&self, f: F) -> T

Calls the given function with the key that was used to intern this value.

This is mainly used to prevent frequent cloning of the key when doing a lookup.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.