ra_salsa
Trait QueryDb

pub trait QueryDb<'d>: Sized {
    type DynDb: ?Sized + Database + HasQueryGroup<Self::Group> + 'd;
    type Group: QueryGroup<GroupStorage = Self::GroupStorage>;
    type GroupStorage;
}

Trait implements by all of the “special types” associated with each of your queries.

Base trait of Query that has a lifetime parameter to allow the DynDb to be non-’static.

Required Associated Types

type DynDb: ?Sized + Database + HasQueryGroup<Self::Group> + 'd

Dyn version of the associated trait for this query group.

type Group: QueryGroup<GroupStorage = Self::GroupStorage>

Associate query group struct.

type GroupStorage

Generated struct that contains storage for all queries in a group.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.